requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
cffi==1.9.1
cryptography==1.7.2
docopt==0.6.2
dominate==2.3.1
enum34==1.1.6
icalendar==3.11.2
idna==2.2
ipaddress==1.0.18
packaging==16.8
pyasn1==0.2.1
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
termcolor==1.1.0
